Fixed issue on 64 bit builds (Fix pointer cast in outputstream)  - ID: 3073281
git-svn-id: https://librecad.svn.sourceforge.net/svnroot/librecad/trunk@61 c6892bbe-2a20-4548-a94f-41916fb2bb95

+++ src/lib/engine/rs_layer.cpp
@@ -55,7 +55,7 @@ std::ostream& operator << (std::ostream& os, const RS_Layer& l) {
     os << " name: " << l.getName().latin1()
     << " pen: " << l.getPen()
 	<< " frozen: " << (int)l.isFrozen()
-	<< " address: " << (int)(&l)
+	<< " address: " << &l
     << std::endl;
     return os;
 }
